# openapi.yaml
openapi: 3.0.0
info:
    title: 'Petrogassa API'
    description: >
        API de autenticación con NestJS (modelo cerrado, sin registro público):
        un admin da de alta usuarios por invitación; cada invitado activa su
        cuenta definiendo nombre, apellido y contraseña. Login con JWT en cookies
        HttpOnly, refresh tokens rotados (hasheados en DB, con detección de
        reutilización) y recuperación de contraseña. Todas las respuestas usan un
        sobre estándar (ver esquemas `SuccessEnvelope` / `ErrorResponse`).
    version: '0.0.1'
servers:
    - url: /api
      description: Servidor local
...
paths:
...
    /news:
        get:
            tags: [News]
            summary: Listar notas publicadas (público, paginado)
            parameters:
                - { in: query, name: category, schema: { type: string, enum: [novedades, prensa] } }
                - { in: query, name: featured, schema: { type: boolean } }
                - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
                - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
            responses:
                '200':
                    description: Notas publicadas, orden publishedAt DESC, con meta de paginación.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/NewsPostEntity'
                                          meta:
                                              $ref: '#/components/schemas/PaginationMeta'
        post:
            tags: [News]
            summary: Crear nota (admin o rrhh)
            security: [{ cookieAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateNewsPostDto'
            responses:
                '201': { description: Nota creada. }
                '400': { $ref: '#/components/responses/ValidationError' }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '409': { $ref: '#/components/responses/UniqueConflict' }
    /news/admin:
        get:
            tags: [News]
            summary: Listar notas incluyendo borradores (admin o rrhh, paginado)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: query, name: category, schema: { type: string, enum: [novedades, prensa] } }
                - { in: query, name: featured, schema: { type: boolean } }
                - { in: query, name: published, schema: { type: boolean } }
                - { in: query, name: page, schema: { type: integer } }
                - { in: query, name: limit, schema: { type: integer } }
            responses:
                '200': { description: Notas obtenidas. }
                '401': { $ref: '#/components/responses/Unauthorized' }
    /news/admin/{id}:
        get:
            tags: [News]
            summary: Obtener una nota por id para edición (admin o rrhh)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Nota obtenida. }
                '404': { $ref: '#/components/responses/NotFound' }
    /news/{slug}:
        get:
            tags: [News]
            summary: Obtener una nota publicada por slug (público)
            parameters:
                - { in: path, name: slug, required: true, schema: { type: string } }
            responses:
                '200': { description: Nota obtenida. }
                '404': { $ref: '#/components/responses/NotFound' }
    /news/{id}:
        patch:
            tags: [News]
            summary: Actualizar nota (admin o rrhh)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Nota actualizada. }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/UniqueConflict' }
        delete:
            tags: [News]
            summary: Enviar nota a la papelera (admin o rrhh, soft delete)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Nota movida a la papelera. }
                '404': { $ref: '#/components/responses/NotFound' }
    /news/admin/trash:
        get:
            tags: [News]
            summary: Papelera de novedades (admin o rrhh)
            security: [{ cookieAuth: [] }]
            responses:
                '200':
                    description: Notas borradas (soft delete), URLs absolutas.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data: { type: array, items: { $ref: '#/components/schemas/NewsPostEntity' } }
                '401': { $ref: '#/components/responses/Unauthorized' }
    /news/{id}/restore:
        post:
            tags: [News]
            summary: Restaurar una nota de la papelera (admin o rrhh)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Nota restaurada. }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: >
                        Sin `conflict`: la nota no está en la papelera. Con
                        `conflict` (inTrash:false): otra nota activa ya usa ese slug
                        (defensivo — la papelera reserva el slug).
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UniqueConflictErrorResponse'
    /news/{id}/permanent:
        delete:
            tags: [News]
            summary: Eliminar una nota definitivamente (admin o rrhh; borra también la portada del almacenamiento)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Nota eliminada definitivamente. }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { description: 'El registro no está en la papelera: hay que enviarlo ahí antes de eliminarlo definitivamente.' }
    /news/linkedin/sync:
        post:
            tags: [News]
            summary: Sincronizar posteos de LinkedIn a la cola (admin o rrhh)
            description: >-
                Trae los posteos recientes de la página de LinkedIn y agrega los
                nuevos como candidatos pendientes. Con LINKEDIN_FETCH_MODE=stub usa
                posteos de ejemplo; con 'api' llama a la Community Management API real
                (requiere credenciales aprobadas). El dedupe por externalId evita
                reimportar lo ya procesado.
            security: [{ cookieAuth: [] }]
            responses:
                '201':
                    description: Sincronización completada.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  fetched: { type: integer, description: Posteos traídos de la fuente }
                                                  created: { type: integer, description: Nuevos agregados a la cola }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '503': { description: LinkedIn no configurado o sin responder (solo modo 'api'). }
    /news/linkedin/imports:
        get:
            tags: [News]
            summary: Cola de curaduría de LinkedIn (admin o rrhh, paginado)
            description: Por defecto lista los candidatos pendientes de aprobación.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: query, name: status, schema: { type: string, enum: [pending, approved, rejected] } }
                - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
                - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 10 } }
            responses:
                '200':
                    description: Candidatos, con meta de paginación.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/LinkedInImportEntity'
                                          meta:
                                              $ref: '#/components/schemas/PaginationMeta'
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
    /news/linkedin/imports/{id}/approve:
        post:
            tags: [News]
            summary: Aprobar y publicar un posteo de LinkedIn (admin o rrhh)
            description: >-
                Crea y publica una nota (news_posts) copiando el contenido del posteo,
                descargando la imagen al almacenamiento, y marca el candidato como aprobado. Devuelve
                la nota creada.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApproveLinkedInImportDto'
            responses:
                '201':
                    description: Nota creada y publicada.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/NewsPostEntity'
                '400': { $ref: '#/components/responses/ValidationError' }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409':
                    description: >
                        Único 409 de este endpoint: el candidato ya fue procesado
                        (aprobado o rechazado). Un slug derivado duplicado NO da
                        409: se le agrega sufijo automáticamente (titulo-2,
                        titulo-3...). No trae `conflict`.
    /news/linkedin/imports/{id}/reject:
        post:
            tags: [News]
            summary: Rechazar un posteo de LinkedIn (admin o rrhh)
            description: Marca el candidato como rechazado; no se publica y no vuelve a la cola.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '201':
                    description: Candidato rechazado.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              $ref: '#/components/schemas/LinkedInImportEntity'
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { description: El candidato ya fue procesado. }
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: accessToken
    responses:
        ValidationError:
            description: Error de validación (Bad Request)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Unauthorized:
            description: No autenticado / credenciales o token inválidos
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Forbidden:
            description: Autenticado pero sin permisos suficientes
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        NotFound:
            description: El recurso solicitado no existe
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        UniqueConflict:
            description: >
                Conflicto por valor único (slug/name) ya tomado. El cuerpo agrega
                `conflict` con el registro que lo retiene y si está en la papelera,
                para poder ofrecer restaurarlo o eliminarlo definitivamente.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/UniqueConflictErrorResponse'
...
    schemas:
        SuccessEnvelope:
            type: object
            description: Sobre estándar de respuestas exitosas.
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: 'Operación exitosa'
                data:
                    description: Carga útil (objeto, array o null).
                meta:
                    $ref: '#/components/schemas/PaginationMeta'
            required: [success, statusCode, message, data]
        PaginationMeta:
            type: object
            description: Metadata de paginación (presente solo en listas paginadas).
            properties:
                totalItems: { type: integer, example: 145 }
                itemsPerPage: { type: integer, example: 10 }
                currentPage: { type: integer, example: 2 }
                totalPages: { type: integer, example: 15 }
                hasNextPage: { type: boolean, example: true }
                hasPrevPage: { type: boolean, example: true }
                nextPageUrl:
                    type: string
                    nullable: true
                    example: '/api/products?limit=10&page=3'
                prevPageUrl:
                    type: string
                    nullable: true
                    example: '/api/products?limit=10&page=1'
        ErrorResponse:
            type: object
            description: Sobre estándar de errores (lo formatean los Exception Filters).
            properties:
                success: { type: boolean, example: false }
                statusCode: { type: integer, example: 400 }
                message:
                    type: string
                    example: 'La contraseña debe tener al menos 8 caracteres...'
                errors:
                    type: array
                    nullable: true
                    items: { type: string }
                timestamp: { type: string, format: date-time }
                path: { type: string, example: '/api/auth/activate' }
        UniqueConflictInfo:
            type: object
            description: >
                Registro que ya usa el valor único en conflicto. Permite al panel
                ofrecer la solución en el momento sin tener que buscarlo.
            properties:
                id: { type: string, format: uuid, description: Id del registro que retiene el valor }
                field: { type: string, enum: [slug, name] }
                value: { type: string, example: 'conflicto-sd' }
                inTrash:
                    type: boolean
                    description: >
                        true = el bloqueante está en la papelera; con su id se puede
                        ofrecer POST /<recurso>/{id}/restore o
                        DELETE /<recurso>/{id}/permanent para liberar el valor.
...
        UniqueConflictErrorResponse:
            description: >
                Sobre de error estándar + `conflict`. Ese campo SOLO aparece en los
                409 por valor único ya tomado; el resto de los errores conserva el
                cuerpo de ErrorResponse.
            allOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - type: object
                  properties:
                      conflict: { $ref: '#/components/schemas/UniqueConflictInfo' }
...
        NewsPostEntity:
            type: object
            properties:
                id: { type: string, format: uuid }
                title: { type: string }
                slug: { type: string }
                category: { type: string, enum: [novedades, prensa] }
                publishedAt: { type: string, format: date-time, nullable: true }
                coverImage: { type: string, nullable: true, description: URL absoluta }
                excerpt: { type: string }
                body: { type: string }
                isFeatured: { type: boolean, description: 'ÚNICA: a lo sumo una nota destacada a la vez (lo garantiza el backend)' }
                isPublished: { type: boolean }
                source: { type: string, enum: [manual, linkedin], description: Origen de la nota }
                externalUrl: { type: string, nullable: true, description: URL del post original en LinkedIn (si fue importada) }
                deletedAt: { type: string, format: date-time, nullable: true, description: 'En la papelera si no es null' }
        CreateNewsPostDto:
            type: object
            required: [title, category, excerpt, body]
            properties:
                title: { type: string, maxLength: 200 }
                slug: { type: string, maxLength: 220, description: 'Opcional; "admin" está reservado. Al crear: se genera del título. Al editar: si no se envía SE CONSERVA (cambiar el título no cambia la URL)' }
                category: { type: string, enum: [novedades, prensa] }
                publishedAt: { type: string, format: date-time, description: Si falta al publicar, se usa la fecha actual }
                coverImage: { type: string, description: 'Key de media (media/AAAA/MM/uuid.ext). Se acepta también la URL absoluta que devuelve el GET: se normaliza a key. Cualquier otro valor da 400.' }
                excerpt: { type: string, maxLength: 500 }
                body: { type: string }
                isFeatured: { type: boolean, description: 'true des-marca automáticamente la destacada anterior (atómico; una sola destacada). false siempre es válido (cero destacadas permitido)' }
                isPublished: { type: boolean, description: false = borrador }
                externalUrl: { type: string }
        LinkedInImportEntity:
            type: object
            description: Candidato de la bandeja de LinkedIn (staging; NO es contenido público).
            properties:
                id: { type: string, format: uuid }
                externalId: { type: string, description: URN del post en LinkedIn }
                status: { type: string, enum: [pending, approved, rejected] }
                externalUrl: { type: string, description: Permalink al post en LinkedIn }
                text: { type: string }
                mediaUrl: { type: string, nullable: true, description: URL de la imagen del post (previsualización) }
                authorName: { type: string, nullable: true }
                postedAt: { type: string, format: date-time, nullable: true }
                newsPostId: { type: string, format: uuid, nullable: true, description: Nota creada al aprobar }
                reviewedAt: { type: string, format: date-time, nullable: true }
        ApproveLinkedInImportDto:
            type: object
            required: [category]
            properties:
                category: { type: string, enum: [novedades, prensa], description: Catalogación (obligatoria) }
                isFeatured: { type: boolean, description: 'true des-marca automáticamente la destacada anterior (una sola destacada)' }
                title: { type: string, maxLength: 200, description: Opcional; si falta se deriva del texto }
                excerpt: { type: string, maxLength: 500, description: Opcional; si falta se deriva del texto }
...